consolidate and restructure
